libflux: create message_private header file
Problem: The message.c file is getting large and we'd like to
split it up.  But there will need to be shared data structures
between files.

Solution: Create a new message_private.h header file that will
contain struct flux_msg.
